moving enums to API module
